# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drawbridge API",
    "version": "0.0.0",
    "description": "The API behind Drawbridge's web UI (docs/PLAN.md §8).\n\nIt's served only to the home network and the VPN, over HTTPS on port 51821. Log in with POST /api/auth/login; the session is a cookie, `__Host-drawbridge`. Every request that changes something needs the `X-Drawbridge` header (any value) and, when it has a body, `Content-Type: application/json`.\n\nA dashboard that can't log in can use a read-only API token instead: make one on the Account page, and send it as `Authorization: Bearer dbt_…`. It works on a short list of GET endpoints (marked below with the `token` security scheme) and is refused everywhere else with 403: it can't read a client's config or DNS log, can't change anything, and can't make or revoke tokens. A token that doesn't exist or was revoked gets 401. It is subject to the same network limits as the web UI."
  },
...
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
...
    "/api/setup": {
      "get": {
        "summary": "Whether first-run setup is needed",
        "responses": {
          "200": {
            "description": "Setup status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatus"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": []
      },
      "post": {
        "summary": "Create the admin account",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "description": "The setup token is wrong, or the request failed the CSRF checks.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "201": {
            "description": "The account exists, and the response logs it in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/SessionCookie"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "409": {
            "description": "Setup is already complete.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Needs the one-time setup token that the package prints at install, and `drawbridge admin setup-token` shows.",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Log in",
        "responses": {
          "200": {
            "description": "Logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/SessionCookie"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "401": {
            "description": "The username or the password is wrong. Or the password is right and the account has two-factor authentication on: with no code, the error's code is totp_required and nothing is counted against the limits; with a wrong or already-used code, it is an ordinary failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "After five failed attempts from a source or for the account, each further attempt waits twice as long as the last, up to 15 minutes. A wrong code counts like a wrong password, and a right password doesn't clear the count, so a code can't be guessed faster than a password can. A code is good once, within a step either side of the current one.",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Log out",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "204": {
            "description": "Logged out; the cookie is cleared."
          }
        },
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "The logged-in account",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "The account and this session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password": {
      "post": {
        "summary": "Change the password",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "204": {
            "description": "Changed. Every other session is logged out."
          },
          "400": {
            "description": "The current password is wrong, or the new one is too short.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          }
        }
      }
    },
    "/api/auth/totp/enroll": {
      "post": {
        "summary": "Start turning on two-factor authentication",
        "description": "Makes a new secret and keeps it, waiting for the first code from the admin's authenticator app (`/api/auth/totp/verify`). It takes the account's password again even in a logged-in session, because whoever turns two-factor authentication on picks the second factor; wrong attempts count against the login limits (429). Nothing changes at login until the first code is proved, and calling it again replaces the secret. The response is the only time the secret is shown. Not available to an API token.",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "The secret to put in the app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TOTPEnrollment"
                }
              }
            }
          },
          "400": {
            "description": "The password is wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TOTPEnrollRequest"
              }
            }
          }
        }
      }
    },
    "/api/auth/totp/verify": {
      "post": {
        "summary": "Finish turning on two-factor authentication",
        "description": "Checks the first code from the app against the secret from `/api/auth/totp/enroll`, and turns two-factor authentication on. From then on a login needs a code. Returns ten recovery codes, **once**: only their hashes are kept. Every other session of the account is logged out, as when the password changes; the one that made this request stays. The code that proves the app is used up. Recorded as an `auth.totp_enabled` event. A wrong code counts against the login limits (429). Not available to an API token.",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "On. The recovery codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodes"
                }
              }
            }
          },
          "400": {
            "description": "The code is wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Two-factor authentication is on already, or no enrollment is waiting: start again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TOTPVerifyRequest"
              }
            }
          }
        }
      }
    },
    "/api/auth/totp/disable": {
      "post": {
        "summary": "Turn off two-factor authentication",
        "description": "Turns it off, and forgets the secret and the recovery codes. It takes the account's password again and a code (from the app, or a recovery code), so a hijacked session can't take the second factor away. Wrong attempts count against the login limits, and a right password doesn't clear the count (429). Every other session of the account is logged out. Recorded as an `auth.totp_disabled` event. From the command line, `drawbridge admin disable-2fa` does it without a code, for an admin who has lost both. Not available to an API token.",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "204": {
            "description": "Off."
          },
          "400": {
            "description": "The password or the code is wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Two-factor authentication isn't on.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecondFactorRequest"
              }
            }
          }
        }
      }
    },
    "/api/auth/totp/recovery-codes": {
      "post": {
        "summary": "Make new recovery codes",
        "description": "Replaces the account's recovery codes with ten new ones, returned **once**; the old ones, used or not, stop working. It takes the account's password again and a code (from the app, or a recovery code), with the same limits as turning two-factor authentication off. Recorded as an `auth.recovery_codes_renewed` event. Not available to an API token.",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "The new codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodes"
                }
              }
            }
          },
          "400": {
            "description": "The password or the code is wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Two-factor authentication isn't on.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecondFactorRequest"
              }
            }
          }
        }
      }
    },
    "/api/auth/sessions": {
      "get": {
        "summary": "Active sessions",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "Newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/sessions/{id}": {
      "delete": {
        "summary": "Revoke a session",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "204": {
            "description": "Revoked. Revoking the current session logs out."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ]
      }
    },
    "/api/server": {
      "get": {
        "summary": "Server settings",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "The settings.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          }
        },
        "tags": [
          "server"
        ]
      },
      "patch": {
        "summary": "Change server settings",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "The saved settings, and whether they're in effect.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Invalid"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "description": "Changes apply to the tunnel at once. Client configs pick up endpoint, port, MTU, DNS, and keepalive changes when they're downloaded again.\n\nA change that could cut the admin off (the listen port, or removing a source from `admin_allowed`; rotating the server's key is `POST /api/server/rotate-key`) is applied on probation: the response carries `pending_change`, and the change is undone after `expires_in` seconds unless `POST /api/server/apply/confirm` keeps it. Until then every other settings change is refused with 409. See `GET /api/server/apply`.",
        "tags": [
          "server"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          }
        }
      }
    },
    "/api/server/status": {
      "get": {
        "summary": "Tunnel status",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "Whether the tunnel is up, and client counts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerStatus"
                }
              }
            }
          }
        },
        "tags": [
          "server"
        ],
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ],
        "description": "Also readable with a read-only API token."
      }
    },
    "/api/server/rotate-key": {
      "post": {
        "summary": "Give the server a new key pair",
        "description": "Replaces the server's private key and public key and applies them to the tunnel at once. Every client's config holds the old public key, so none of them can connect until it imports the new config; the clients that had been handed a config show `config_outdated`. The rotation is applied on probation, like any change that could cut the admin off: the response carries `pending_change`, and the old key is restored after `expires_in` seconds unless `POST /api/server/apply/confirm` keeps it. Until then every other settings change is refused with 409. The event is `server.key_rotated`.",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "Rotated, and waiting to be kept. The settings as they are now, with the new `public_key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResult"
                }
              }
            }
          },
          "409": {
            "description": "Another settings change is waiting to be kept or undone.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "server"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ]
      }
    },
    "/api/server/apply": {
      "get": {
        "summary": "Is a settings change waiting to be kept?",
        "description": "A settings change that could cut the admin off is applied on probation and undone unless it's kept in time (docs/PLAN.md §4.3). This says whether one is waiting. The same is pushed in the `status` message of `GET /api/stream`. Not readable with an API token.",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "The change waiting, or an empty object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyState"
                }
              }
            }
          }
        },
        "tags": [
          "server"
        ]
      }
    },
    "/api/server/apply/confirm": {
      "post": {
        "summary": "Keep the settings change that is waiting",
        "description": "Keeps the change, so it isn't undone. A request that arrives after the deadline doesn't keep it: the change is undone and the answer is 409.",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "Kept. The settings as they are now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResult"
                }
              }
            }
          },
          "409": {
            "description": "No change is waiting (it was already kept or undone), or its time was up and it was undone.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "server"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ]
      }
    },
    "/api/server/apply/revert": {
      "post": {
        "summary": "Undo the settings change that is waiting, now",
        "description": "Restores the settings as they were before the change and applies them, without waiting for the time to run out.",
        "responses": {
          "401": {
            "$ref": "#/components/responses/NotLoggedIn"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "200": {
            "description": "Undone. The settings as they are now.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResult"
                }
              }
            }
          },
          "409": {
            "description": "No change is waiting.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "server"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CSRF"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SetupStatus": {
        "type": "object",
        "required": [
          "needed"
        ],
        "properties": {
          "needed": {
            "type": "boolean",
            "description": "True until the admin account exists."
          }
        }
      },
...
      "Me": {
        "type": "object",
        "required": [
          "user",
          "session"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-drawbridge"
      },
      "token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "dbt_ and 43 characters",
        "description": "A read-only API token (docs/PLAN.md §6.5), made on the Account page and shown once. Accepted only by the operations that list it; any other gets 403."
      }
    },
...
    "responses": {
...
      "NotLoggedIn": {
        "description": "Not logged in, or the session ended.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}